+++ src/preprocessor.ts
@@ -1,12 +1,18 @@
+// TODO: give macros the ability to have arguments and call other macros
+
 import * as fs from 'fs';
-const { readdir, readFile, writeFile, mkdir, access } = fs.promises;
+import { workspace } from 'vscode';
+import { sep } from 'path';
+import { normalize } from 'path';
+const { readdir, readFile, writeFile, mkdir, unlink, access } = fs.promises;
 
 export const preprocess = async (path: string) => {
   const macros: {
     [key: string]: {
       text: string,
       lets: {[key: string]: string},
-      constants: {[key: string]: string}
+      constants: {[key: string]: string},
+      arguments: string[]
     }
   } = {};
   const outFiles: {[key: string]: string} = {};
@@ -38,7 +44,23 @@ export const preprocess = async (path: string) => {
     updateScopeItems();
   };
 
-  const processLine = (line: string, file: string, lineNum: number) => {
+  const convertRawVars = (line: string) => {
+    for (const [alias, variable] of Object.entries(currentScopeItems.lets)) {
+      line = line.replace(new RegExp(`\\b${alias}\\b`, 'g'), variable);
+    }
+    for (const [alias, value] of Object.entries(currentScopeItems.constants)) {
+      line = line.replace(new RegExp(`\\b${alias}\\b`, 'g'), value);
+    }
+    for (const [alias, variable] of Object.entries(globals.lets)) {
+      line = line.replace(new RegExp(`\\b${alias}\\b`, 'g'), variable);
+    }
+    for (const [alias, value] of Object.entries(globals.constants)) {
+      line = line.replace(new RegExp(`\\b${alias}\\b`, 'g'), value);
+    }
+    return line;
+  }
+
+  const processLine = (line: string, file: string, lineNum: number, localSnippets?: typeof macros): string => {
     const errStart = errStartGen(file, lineNum);
     let isError = false;
     // scope stack modifiers
@@ -58,10 +80,10 @@ export const preprocess = async (path: string) => {
     // these "return" so as not to add to the actual output
     else if (line.trim().startsWith("#macro")) {
       errors.push(`${errStart}macros must be defined within macros.as;\nmacros cannot be defined within a macro`);
-      return;
+      return "error";
     }
     else if (line.trim().startsWith("#let")) {
-      const matches = line.trim().match(/#let\s+(?<alias>[a-zA-Z_]+)\s*=\s*(?<variable>var[0-9]+)/);
+      const matches = line.trim().match(/#let\s+(?<alias>[a-zA-Z_][a-zA-Z0-9_]+)\s*=\s*(?<variable>var[0-9]+)/);
       if (!matches!.groups!.alias) {
         errors.push(`${errStart}let-variable must have a name`);
       }
@@ -71,10 +93,10 @@ export const preprocess = async (path: string) => {
       if (isError) { return "error"; }
       if (file === "globals.as") { globals.lets[matches!.groups!.alias] = matches!.groups!.variable; }
       else { addLetToScope(matches!.groups!); }
-      return;
+      return "";
     }
     else if (line.trim().startsWith("#const")) {
-      const matches = line.trim().match(/#const\s+(?<alias>[a-zA-Z_]+)\s*=\s*(?<value>[0-9]+(?:\.[0-9]+)?)/);
+      const matches = line.trim().match(/#const\s+(?<alias>[a-zA-Z_][a-zA-Z0-9_]+)\s*=\s*(?<value>-?[0-9]+(?:\.[0-9]+)?)/);
       if (!matches!.groups!.alias) {
         errors.push(`${errStart}const must have a name`);
       }
@@ -84,18 +106,71 @@ export const preprocess = async (path: string) => {
       if (isError) { return "error"; }
       if (file === "globals.as") { globals.constants[matches!.groups!.alias] = matches!.groups!.value; }
       else { addConstToScope(matches!.groups!); }
-      return;
+      return "";
+    }
+    if (localSnippets) {
+      for (const [snippetName, parsedSnippet] of Object.entries(localSnippets)) {
+        if (line.trim().startsWith("{" + snippetName + "}")) {  
+          line = parsedSnippet.text;
+  
+          const procLines: (string | undefined)[] = [];
+  
+          scopeStack.push({lets: {}, constants: {}});
+          for (const [ind, ln] of line.split("\n").entries()) {
+            const processed = processLine(ln, `(${errStart}) > ${snippetName}`, ind);
+            if (processed) { procLines.push(processed); }
+          }
+          const parsedSnippetItems = scopeStack.pop();
+  
+          const ssTarget = scopeStack[scopeStack.length - 1];
+          ssTarget.lets = { ...ssTarget.lets, ...parsedSnippetItems!.lets };
+          ssTarget.constants = { ...ssTarget.constants, ...parsedSnippetItems!.constants };
+          updateScopeItems();
+  
+          if (!file.startsWith("(")) {
+            if (!outFiles[file]) { outFiles[file] = ''; }
+            outFiles[file] += `${procLines.join('\n')}\n`;
+          }
+          return procLines.join('\n');
+        }
+      }
     }
     for (const [macroName, parsedMacro] of Object.entries(macros)) {
-      if (line.trim() === macroName) {
+      if (line.trim().startsWith(macroName)) {
+        const passedArgRegexRes = /\((?<args>.*)\)/.exec(line);
+        const passedArgs = (passedArgRegexRes?.groups?.args) ? passedArgRegexRes.groups.args.split(",").map(arg=>arg.trim()) : [];
+
+        if (passedArgs.length !== parsedMacro.arguments.length) {
+          errors.push(`${errStart}macro argument count mismatch!`);
+          return "error";
+        }
+
+        const processedArgs: string[] = passedArgs.map(arg => convertRawVars(arg));
+
         line = parsedMacro.text;
+        for (const [idx, arg] of parsedMacro.arguments.entries()) {
+          line = line.replace(new RegExp(`\\{${arg}\\}`, 'g'), processedArgs[idx]);
+        }
+
+        const procLines: (string | undefined)[] = [];
+
+        scopeStack.push({lets: {}, constants: {}});
+        for (const [ind, ln] of line.split("\n").entries()) {
+          const processed = processLine(ln, `(${errStart}) > ${macroName}`, ind);
+          if (processed) { procLines.push(processed); }
+        }
+        const parsedMacroItems = scopeStack.pop();
+
         const ssTarget = scopeStack[scopeStack.length - 1];
-        ssTarget.lets = { ...ssTarget.lets, ...parsedMacro.lets };
-        ssTarget.constants = { ...ssTarget.constants, ...parsedMacro.constants };
+        ssTarget.lets = { ...ssTarget.lets, ...parsedMacroItems!.lets };
+        ssTarget.constants = { ...ssTarget.constants, ...parsedMacroItems!.constants };
         updateScopeItems();
-        if (!outFiles[file]) { outFiles[file] = ''; }
-        outFiles[file] += `${line}\n`;
-        return line;
+
+        if (!file.startsWith("(")) {
+          if (!outFiles[file]) { outFiles[file] = ''; }
+          outFiles[file] += `${procLines.join('\n')}\n`;
+        }
+        return procLines.join('\n');
       }
     }
 
@@ -148,10 +223,69 @@ export const preprocess = async (path: string) => {
         return out.join(" ");
       });
 
-    if (!outFiles[file]) { outFiles[file] = ''; }
-    outFiles[file] += `${line}\n`;
+    if (!file.startsWith("(")) {
+      if (!outFiles[file]) { outFiles[file] = ''; }
+      outFiles[file] += `${line}\n`;
+    }
     return line;
   };
+
+  const includePath: string = workspace.getConfiguration("aiscriptpad").get("includepath") as string;
+  let sharedPath = includePath;
+  if (includePath.startsWith('.')) {
+    workspace.workspaceFolders?.map((folder) => {
+      const fPath = folder.uri.path;
+      if (fs.readdirSync(fPath.substring(sep === "\\" ? fPath.indexOf("/") + 1 : 0)).includes(includePath.substring(2))) {
+        sharedPath = folder.uri.path + includePath.substring(1);
+      }
+    });
+  }
+  if (sep === '\\') {
+    sharedPath = sharedPath.substring(sharedPath.indexOf('/') + 1);
+  }
+
+  let templates: string[] = [];
+  
+  if (await (await readdir(`${sharedPath}`)).includes('shared')) {
+    const files = await readdir(`${sharedPath}/shared`);
+    if (files.includes('templates')) {
+      templates = await readdir(`${sharedPath}/shared/templates`);
+    }
+    if (files.includes('globals.as')) {
+      const lines = (await readFile(`${sharedPath}/shared/globals.as`, 'utf8')).split(/\r?\n/g);
+      for (const [lineNum, line] of lines.entries()) {
+        processLine(line.trim(), "globals.as", lineNum);
+      }
+    }
+    if (files.includes("macros.as")) {
+      const lines = (await readFile(`${sharedPath}/shared/macros.as`, 'utf8')).split(/\r?\n/g);
+      let currentMacroName: undefined | string = undefined;
+      let currentMacroSpaces: undefined | string = undefined;
+      for (const [lineNum, line] of lines.entries()) {
+        if (line.trim().startsWith("#macro")) {
+          const matches = line.match(/^(?<spaces>\s*)#macro\s+(?<name>[a-zA-Z_]+)\((?<args>.*)\)/)!;
+          if (matches.groups!.name) {
+            currentMacroName = matches.groups!.name;
+            macros[matches.groups!.name] = {
+              text: '',
+              lets: {},
+              constants: {},
+              arguments: matches.groups!.args.split(",").map(arg=>arg.trim()).filter(arg=>arg.length > 0),
+            };
+            if (matches.groups!.spaces || matches.groups!.spaces.length === 0) {
+              currentMacroSpaces = matches.groups!.spaces;
+            }
+          }
+        } else if (currentMacroName && line.trim().startsWith("#endmacro")) {
+          currentMacroName = undefined;
+          currentMacroSpaces = undefined;
+        } else if (currentMacroName) {
+          const contentLine = line.substring(currentMacroSpaces!.length);
+          macros[currentMacroName].text += contentLine + '\n';
+        }
+      }
+    }
+  }
 
   const files = await readdir(`${path}`);
 
@@ -167,33 +301,25 @@ export const preprocess = async (path: string) => {
     let currentMacroSpaces: undefined | string = undefined;
     for (const [lineNum, line] of lines.entries()) {
       if (line.trim().startsWith("#macro")) {
-        const matches = line.match(/^(?<spaces>\s*)#macro\s+(?<name>[a-zA-Z_]+)/)!;
+        const matches = line.match(/^(?<spaces>\s*)#macro\s+(?<name>[a-zA-Z_]+)\((?<args>.*)\)/)!;
         if (matches.groups!.name) {
           currentMacroName = matches.groups!.name;
           macros[matches.groups!.name] = {
             text: '',
             lets: {},
             constants: {},
+            arguments: matches.groups!.args.split(",").map(arg=>arg.trim()).filter(arg=>arg.length > 0),
           };
           if (matches.groups!.spaces || matches.groups!.spaces.length === 0) {
             currentMacroSpaces = matches.groups!.spaces;
           }
-          scopeStack.push({
-            lets: {},
-            constants: {}
-          });
         }
       } else if (currentMacroName && line.trim().startsWith("#endmacro")) {
-        macros[currentMacroName].lets = currentScopeItems.lets;
-        macros[currentMacroName].constants = currentScopeItems.constants;
-        scopeStack.pop();
-        updateScopeItems();
         currentMacroName = undefined;
         currentMacroSpaces = undefined;
       } else if (currentMacroName) {
         const contentLine = line.substring(currentMacroSpaces!.length);
-        const processedLine = processLine(contentLine, "macros.as", lineNum);
-        if (processedLine) { macros[currentMacroName].text += processedLine + '\n'; }
+        macros[currentMacroName].text += contentLine + '\n';
       }
     }
   }
@@ -206,8 +332,62 @@ export const preprocess = async (path: string) => {
       lets: {},
       constants: {}
     });
+
+    if (!lines.some(ln => ln.trim().startsWith("id ")) && !["AIPD.aipd", "ATKD.atkd"].includes(file)) {
+      if (!templates.includes(file)) {
+        errors.push(`no template found for ${file}!`);
+        continue;
+      }
+
+      templates.splice(templates.indexOf(file), 1);
+      const templateLines = (await readFile(`${sharedPath}/shared/templates/${file}`, 'utf8')).split(/\r?\n/g);
+      const snippets: typeof macros = {};
+      let currentSnippetName: undefined | string = undefined;
+      let currentSnippetSpaces: undefined | string = undefined;
+      for (const [lineNum, line] of lines.entries()) {
+        if (line.trim().startsWith("#snippet")) {
+          const matches = line.match(/^(?<spaces>\s*)#snippet\s+(?<name>[a-zA-Z_]+)/)!;
+          if (matches.groups!.name) {
+            currentSnippetName = matches.groups!.name;
+            snippets[matches.groups!.name] = {
+              text: '',
+              lets: {},
+              constants: {},
+              arguments: [],
+            };
+            if (matches.groups!.spaces || matches.groups!.spaces.length === 0) {
+              currentSnippetSpaces = matches.groups!.spaces;
+            }
+          }
+        } else if (currentSnippetName && line.trim().startsWith("#endsnippet")) {
+          currentSnippetName = undefined;
+          currentSnippetSpaces = undefined;
+        } else if (currentSnippetName) {
+          const contentLine = line.substring(currentSnippetSpaces!.length);
+          snippets[currentSnippetName].text += contentLine + '\n';
+        }
+      }
+
+      for (const [lineNum, line] of templateLines.entries()) {
+        processLine(line, file, lineNum, snippets);
+      }
+    } else {
+      for (const [lineNum, line] of lines.entries()) {
+        processLine(line, file, lineNum);
+      }
+    }
+    scopeStack.pop();
+  }
+
+  for (const template of templates) {
+    const lines = (await readFile(`${sharedPath}/shared/templates/${template}`, 'utf8')).split(/\r?\n/g);
+
+    scopeStack.push({
+      lets: {},
+      constants: {}
+    });
     for (const [lineNum, line] of lines.entries()) {
-      processLine(line, file, lineNum);
+      processLine(line, template, lineNum);
     }
     scopeStack.pop();
   }
@@ -216,7 +396,15 @@ export const preprocess = async (path: string) => {
     return errors;
   }
 
-  if (!fs.existsSync(`${path}/__preprocessed`)) { await mkdir(`${path}/__preprocessed`); }
+  if (fs.existsSync(`${path}/__preprocessed`)) { 
+    const inPreprocessed = await readdir(`${path}/__preprocessed`);
+    for (const file of inPreprocessed) {
+      if (!Object.keys(outFiles).includes(file)) await unlink(`${path}/__preprocessed/${file}`);
+    }
+  } else {
+    await mkdir(`${path}/__preprocessed`);
+  }
+  
 
   for (const [name, content] of Object.entries(outFiles)) {
     if (["globals.as", "macros.as", "__preprocessed"].includes(name)) { continue; }